fix: conflicts, add merged tests class
