+++ inlong-sdk/transform-sdk/src/test/java/org/apache/inlong/sdk/transform/process/function/string/TestStringFunctionsProcessor.java
@@ -28,6 +28,8 @@ import org.junit.runners.Suite;
         TestLocateFunction.class, TestLowerFunction.class, TestLpadFunction.class, TestReplaceFunction.class,
         TestReplicateFunction.class, TestReverseFunction.class, TestRightFunction.class, TestRpadFunction.class,
         TestSpaceFunction.class, TestStrcmpFunction.class, TestSubstringFunction.class, TestToBase64Function.class,
-        TestTranslateFunction.class, TestTrimFunction.class, TestUpperFunction.class})
+        TestTranslateFunction.class, TestTrimFunction.class, TestUpperFunction.class, TestContainsFunction.class,
+        TestFromBase64Function.class
+})
 public class TestStringFunctionsProcessor {
 }
